refactor: extract for reuse
