+++ scriptUtils.js
@@ -1,6 +1,44 @@
 const path = require('path');
 const fs = require('node:fs');
 const { globSync } = require('glob');
+const { pathPrefix: pathPrefixFromGatsbyConfig } = require('./gatsby-config.js');
+
+function getPathPrefixFromConfig() {
+    const CONFIG_PATH = path.join('src', 'pages', 'config.md');
+    if (!fs.existsSync(CONFIG_PATH)) {
+        return null;
+    }
+
+    const data = fs.readFileSync(CONFIG_PATH).toString();
+    if (!data) {
+        return null;
+    }
+
+    const lines = data.split('\n');
+
+    // find the pathPrefix key
+    const keyIndex = lines.findIndex((line) => new RegExp(/\s*-\s*pathPrefix:/).test(line));
+    if (keyIndex < 0) {
+        return null;
+    }
+
+    // find the pathPrefix value
+    const line = lines.slice(keyIndex + 1)?.find((line) => new RegExp(/\s*-/).test(line));
+    if (!line) {
+        null;
+    }
+
+    // extract pathPrefix
+    const pathPrefixLine = line.match(new RegExp(/(\s*-\s*)(\S*)(\s*)/));
+    if (!pathPrefixLine) {
+        return null;
+    }
+    return pathPrefixLine[2];
+}
+
+function getPathPrefix() {
+    return getPathPrefixFromConfig() ?? pathPrefixFromGatsbyConfig;
+}
 
 function getRedirectionsFilePath() {
     const redirectionsFilePath = path.join(__dirname, 'src', 'pages', 'redirects.json');
@@ -61,6 +99,7 @@ function replaceLinksInFile({ file, linkMap, getFindPattern, getReplacePattern }
 }
 
 module.exports = {
+    getPathPrefix,
     getRedirectionsFilePath,
     readRedirectionsFile,
     writeRedirectionsFile,
